requirements.txt
streamlit
pypdf
python-pptx
scikit-learn
langchain-community
langchain-openai
openai
python-dotenv
chromadb
numpy
sentence-transformers
